feat: implement CRUD for menus module and add RestaurantMenu model
re #28
