--- src/modules/menus/menus.module.ts
@@ -1,10 +1,0 @@
-import { Module } from "@nestjs/common";
-import { MenusController } from './controllers/menus.controller';
-import { MenusService } from './services/menus.service';
-import { PrismaService } from "src/prisma/prisma.service";
-
-@Module({
-    controllers: [MenusController],
-    providers: [MenusService, PrismaService]
-})
-export class MenusModule {}
